# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAPI API Documentation",
        "description": "Documentación de la API REST del proyecto SWAPI",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server Local"
        }
    ],
    "paths": {
...
        "/api/character/{id}": {
            "get": {
                "tags": [
                    "SWAPI"
                ],
                "summary": "Obtener y almacenar un personaje",
                "operationId": "4d59791c46b91004babf625d5ae7cf05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del personaje en SWAPI",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Personaje almacenado y retornado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personaje no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/planet/{id}": {
            "get": {
                "tags": [
                    "SWAPI"
                ],
                "summary": "Obtener y almacenar un planeta",
                "operationId": "7435507c60b2733153466a1ca2411df1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del planeta en SWAPI",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Planeta almacenado y retornado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Planeta no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/film/{id}": {
            "get": {
                "tags": [
                    "SWAPI"
                ],
                "summary": "Obtener y almacenar una película",
                "operationId": "824ea3e65d70966046e151348435385e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la película en SWAPI",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Película almacenada y retornada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Película no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}